Basic Chatbot Application/requirements.txt
fastapi
uvicorn
vertexai
python-dotenv